Project2/requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
pyyaml
tqdm
Pillow
seaborn
wandb